+++ api/fe.js
@@ -1,0 +1,6 @@
+const { generateScript, generateContainerScript } = require('../forward_engineering/api');
+
+module.exports = {
+	generateScript,
+	generateContainerScript,
+};
