+++ forward_engineering/api.js
@@ -1,11 +1,7 @@
 const _ = require('lodash');
-const { generateCollectionScript } = require('./services/protoScriptGenerationService');
 const { formatComment } = require('./helpers/utils');
-
-const RECORD_NAME_STRATEGY = 'RecordNameStrategy';
-const TOPIC_RECORD_NAME_STRATEGY = 'TopicRecordNameStrategy';
-const protobufjs = require('protobufjs');
-const descriptor = require('protobufjs/ext/descriptor');
+const { prepareScript } = require('./helpers/prepareScript');
+const { generateCollectionScript } = require('./services/protoScriptGenerationService');
 
 const defaultContainerData = [
 	{
@@ -66,7 +62,7 @@ module.exports = {
 			]
 				.filter(row => row !== '')
 				.join('\n');
-			callback(null, this.prepareScript(script, preparedData));
+			callback(null, prepareScript(script, preparedData));
 		} catch (error) {
 			const errorObject = {
 				message: error.message,
@@ -100,7 +96,7 @@ module.exports = {
 			]
 				.filter(row => row !== '')
 				.join('\n');
-			callback(null, this.prepareScript(script, preparedData));
+			callback(null, prepareScript(script, preparedData));
 		} catch (error) {
 			const errorObject = {
 				message: error.message,
@@ -109,73 +105,5 @@ module.exports = {
 			logger.log('error', errorObject, 'Protobuf Forward-Engineering Error');
 			callback(errorObject);
 		}
-	},
-
-	prepareScript(script, data) {
-		const targetSchemaRegistry = _.get(data, 'options.targetScriptOptions.keyword');
-		if (targetSchemaRegistry === 'confluentSchemaRegistry') {
-			return this.getConfluentPostQuery({ data, schema: script });
-		}
-		if (targetSchemaRegistry === 'pulsarSchemaRegistry') {
-			return this.getPulsarPostQuery({ data, schema: script });
-		}
-
-		return script;
-	},
-
-	getPulsarPostQuery({ data, schema }) {
-		const root = protobufjs.parse(schema).root;
-		const descriptorMsg = root.toDescriptor('proto3');
-		const buffer = descriptor.FileDescriptorSet.encode(descriptorMsg).finish();
-		const fileDescriptorSet = buffer.toString('base64');
-		const descriptorJson = descriptorMsg.toJSON();
-		const rootMessageTypeName = `${_.get(descriptorJson, 'file[0].package')}.${_.get(descriptorJson, 'file[0].messageType[0].name')}`;
-		const rootFileDescriptorName = _.get(descriptorJson, 'file[0].name');
-		const body = {
-			fileDescriptorSet,
-			rootMessageTypeName,
-			rootFileDescriptorName,
-		};
-		const bodyObject = {
-			type: 'PROTOBUF_NATIVE',
-			data: body,
-			properties: {},
-		};
-		const namespace = _.get(data, 'containerData[0].name', '');
-		const topic = _.get(data, 'containerData[0].pulsarTopicName', '');
-		const persistence = _.get(data, 'containerData[0].isNonPersistentTopic', false)
-			? 'non-persistent'
-			: 'persistent';
-		return `POST /${persistence}/${namespace}/${topic}/schema\n\n${JSON.stringify(bodyObject, null, 4)}`;
-	},
-
-	getConfluentPostQuery({ data, schema }) {
-		const getName = () => {
-			const name = this.getRecordName(data);
-
-			const schemaType = _.get(data, 'containerData[0].schemaType');
-			const containerName = _.get(data, 'containerData[0].name');
-			const topic = _.get(data, 'modelData[0].schemaTopic');
-
-			const typePostfix = schemaType ? `-${schemaType}` : '';
-			const containerPrefix = containerName ? `${containerName}.` : '';
-			const topicPrefix = topic ? `${topic}-` : '';
-
-			const schemaNameStrategy = _.get(data, 'modelData[0].schemaNameStrategy', '');
-			switch (schemaNameStrategy) {
-				case RECORD_NAME_STRATEGY:
-					return `${containerPrefix}${name}${typePostfix}`;
-				case TOPIC_RECORD_NAME_STRATEGY:
-					return `${topicPrefix}${containerPrefix}${name}${typePostfix}`;
-				default:
-					return `${name}${typePostfix}`;
-			}
-		};
-
-		return `POST /subjects/${getName()}/versions\n\n${schema}`;
-	},
-
-	getRecordName(data) {
-		return data.containerData[0].code || data.containerData[0].name || data.containerData[0].collectionName;
 	},
 };
